cli: splitting code
